+++ lib/activerecord-pg-format-db-structure/deparser.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
 require "pg_query"
+require "anbt-sql-formatter/formatter"
 
 module ActiveRecordPgFormatDbStructure
   # Returns a list of SQL strings from a list of PgQuery::RawStmt.
   class Deparser
     attr_reader :source
+
+    PRETTY_INDENT_STRING = "    "
 
     def initialize(source)
       @source = source
@@ -19,17 +22,29 @@ module ActiveRecordPgFormatDbStructure
         deparse_index_stmt(raw_statement.stmt.index_stmt)
       in stmt: { alter_table_stmt: _ }
         deparse_alter_table_stmt(raw_statement.stmt.alter_table_stmt)
+      in stmt: { select_stmt: _ }
+        deparse_query_statement(raw_statement.stmt.select_stmt)
+      in stmt: { insert_stmt: _ }
+        deparse_insert_statement(raw_statement.stmt.insert_stmt)
+      in stmt: { create_table_as_stmt: _ }
+        deparse_create_table_as_stmt(raw_statement.stmt.create_table_as_stmt)
+      in stmt: { view_stmt: _ }
+        deparse_view_stmt(raw_statement.stmt.view_stmt)
       else
-        keep_original_string(raw_statement)
+        "\n#{deparse_stmt(raw_statement.stmt.inner)}"
       end
     end
 
     private
 
-    def keep_original_string(raw_statement)
-      start = raw_statement.stmt_location || 0
-      stop = start + raw_statement.stmt_len
-      source[start..stop]
+    def deparse_stmt(stmt)
+      "\n#{PgQuery.deparse_stmt(stmt)};"
+    end
+
+    def deparse_query_statement(stmt)
+      generic_query_str = +"\n\n"
+      generic_query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt))
+      generic_query_str << ";"
     end
 
     def deparse_index_stmt(index_stmt)
@@ -37,40 +52,134 @@ module ActiveRecordPgFormatDbStructure
     end
 
     def deparse_alter_table_stmt(alter_table_stmt)
-      "\n#{
-        deparse_stmt(alter_table_stmt)
-          .gsub(" ADD ", "\n  ADD ")
-          .gsub(" ALTER COLUMN ", "\n  ALTER COLUMN ")
-      }"
+      alter_table_str = +"\n\n"
+      alter_table_str << PgQuery.deparse_stmt(
+        PgQuery::AlterTableStmt.new(
+          **alter_table_stmt.to_h,
+          cmds: []
+        )
+      ).chomp(" ")
+
+      alter_table_cmds_str = alter_table_stmt.cmds.map do |cmd|
+        "\n  #{deparse_alter_table_cmd(cmd)}"
+      end.join(",")
+
+      alter_table_str << alter_table_cmds_str
+      alter_table_str << ";"
+      alter_table_str
     end
 
-    def deparse_stmt(stmt)
-      "\n#{PgQuery.deparse_stmt(stmt)};"
+    def deparse_alter_table_cmd(cmd)
+      PgQuery.deparse_stmt(
+        PgQuery::AlterTableStmt.new(
+          relation: { relname: "tmp" },
+          cmds: [cmd]
+        )
+      ).gsub("ALTER ONLY tmp ", "")
     end
 
     def deparse_create_stmt(create_stmt)
+      placeholder_column = PgQuery::Node.from(
+        PgQuery::ColumnDef.new(
+          colname: "placeholder_column",
+          type_name: {
+            names: [PgQuery::Node.from_string("placeholder_type")]
+          }
+        )
+      )
+
       table_str = "\n\n\n-- Name: #{create_stmt.relation.relname}; Type: TABLE;\n\n"
       table_str << PgQuery.deparse_stmt(
         PgQuery::CreateStmt.new(
           **create_stmt.to_h,
-          table_elts: []
+          table_elts: [placeholder_column]
         )
       )
-      table_str.gsub!(/\(\)\z/, "")
-      table_str << "("
-      table_str << create_stmt.table_elts.map do |elt|
+      table_str << ";"
+
+      table_columns = create_stmt.table_elts.map do |elt|
         "\n    #{deparse_table_elt(elt)}"
       end.join(",")
-      table_str << "\n);"
+      table_columns << "\n"
+
+      table_str[deparse_table_elt(placeholder_column)] = table_columns
+
       table_str
     end
 
     def deparse_table_elt(elt)
       PgQuery.deparse_stmt(
         PgQuery::CreateStmt.new(
-          relation: PgQuery::RangeVar.new(relname: "tmp"), table_elts: [elt]
+          relation: { relname: "tmp" }, table_elts: [elt]
         )
-      ).sub(/\ACREATE TABLE ONLY tmp \(/, "").sub(/\)\z/, "")
+      ).gsub(/\ACREATE TABLE ONLY tmp \((.*)\)\z/, '\1')
+    end
+
+    def deparse_create_table_as_stmt(stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      create_table_as_stmt_str = +"\n\n"
+      create_table_as_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::CreateTableAsStmt.new(
+          **stmt.to_h,
+          query: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      create_table_as_stmt_str << ";"
+
+      query_str = +"(\n"
+      query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt.query.inner)).gsub(/^/, PRETTY_INDENT_STRING)
+      query_str << "\n)"
+
+      create_table_as_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      create_table_as_stmt_str
+    end
+
+    def deparse_view_stmt(stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      view_stmt_str = +"\n\n"
+      view_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::ViewStmt.new(
+          **stmt.to_h,
+          query: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      view_stmt_str << ";"
+
+      query_str = +"(\n"
+      query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt.query.inner)).gsub(/^/, PRETTY_INDENT_STRING)
+      query_str << "\n)"
+
+      view_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      view_stmt_str
+    end
+
+    def deparse_insert_statement(insert_stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      insert_stmt_str = +"\n\n\n"
+      insert_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::InsertStmt.new(
+          **insert_stmt.to_h,
+          select_stmt: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      insert_stmt_str << "\n;"
+
+      query_str = pretty_formt_sql_string(PgQuery.deparse_stmt(insert_stmt.select_stmt.inner))
+      query_str.gsub!("VALUES (", "VALUES\n (")
+
+      insert_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      insert_stmt_str
+    end
+
+    def pretty_formt_sql_string(sql)
+      rule = AnbtSql::Rule.new
+      rule.keyword = AnbtSql::Rule::KEYWORD_UPPER_CASE
+      rule.indent_string = PRETTY_INDENT_STRING
+      formatter = AnbtSql::Formatter.new(rule)
+      formatter.format(sql)
     end
   end
 end