+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.1.1] - pending
+
+- Proper deparsing of all statements
+- Formatting of select and insert sub-statements using `anbt-sql-formatter`
+
 ## [0.1.0] - 2025-01-24
 
 - Initial release